gotham::handler

Trait IntoResponse

pub trait IntoResponse {
    // Required method
    fn into_response(self, state: &State) -> Response<Body>;
}

Represents a type which can be converted to a response. This trait is used in converting the return type of a function into a response.

Examples

struct MyStruct {
    value: String,
}

impl MyStruct {
    fn new() -> MyStruct {
        // ...
    }
}

impl IntoResponse for MyStruct {
    fn into_response(self, _state: &State) -> Response<Body> {
        Response::builder()
            .status(StatusCode::OK)
            .body(self.value.into())
            .unwrap()
    }
}

fn handler(state: State) -> (State, MyStruct) {
    (state, MyStruct::new())
}

gotham::start("127.0.0.1:7878", || Ok(handler)).unwrap();

Required Methods

fn into_response(self, state: &State) -> Response<Body>

Converts this value into a hyper::Response