gotham::handler

Type Alias HandlerFuture

pub type HandlerFuture = dyn Future<Output = HandlerResult> + Send;

A type alias for the trait objects returned by HandlerService.

When the Future resolves to an error, the (State, HandlerError) value is used to generate an appropriate HTTP error response.