gotham

Module router

Defines the Gotham Router and supporting types.

Re-exports

pub use builder::build_router;

pub use builder::build_simple_router;

Modules

builder

Defines a builder API for constructing a Router.

response

Defines Router functionality which acts on the Response

route

Defines types that support individual application routes.

tree

Defines a hierarchial Tree with subtrees of Node.

Structs

RouteNonMatch

The error type used for a non-matching route, as returned by RouteMatcher::is_match. Multiple values of this type can be combined by matchers that are wrapping other matchers, using the intersection / union methods. The data within is used by the Router to create a Response when no routes were successfully matched.

Router

Responsible for dispatching HTTP requests to defined routes, and responding with appropriate error codes when a valid Route is unable to be determined or the dispatch cannot be performed.