gotham::router

Module response

Defines Router functionality which acts on the Response

Structs

NoopResponseExtender

An extender that does not alter the response.

ResponseFinalizer

Holds an immutable collection of ResponseExtender values, as configured using ResponseFinalizerBuilder::add. This type is constructed automatically when using the gotham::router::builder API. See RouterBuilder::add_response_extender for details on configuring ResponseExtender values for each StatusCode.

ResponseFinalizerBuilder

Builds an immutable ResponseFinalizer.

Traits

ResponseExtender

Allow complex types to extend the Response based on current State and Response data.

StaticResponseExtender

Extend the Response based on current State and Response data.

Derive Macros

StaticResponseExtender