gotham::router::route::matcher

Trait RouteMatcher

pub trait RouteMatcher: RefUnwindSafe + Clone {
    // Required method
    fn is_match(&self, state: &State) -> Result<(), RouteNonMatch>;
}

Determines if conditions required for the associated Route to be invoked by the Router have been met.

Required Methods

fn is_match(&self, state: &State) -> Result<(), RouteNonMatch>

Determines if the Request meets pre-defined conditions.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.