gotham::router

Struct Router

pub struct Router { /* private fields */ }

Responsible for dispatching HTTP requests to defined routes, and responding with appropriate error codes when a valid Route is unable to be determined or the dispatch cannot be performed.

A Router is constructed through the gotham::router::builder API, and used with the gotham::start function when booting a Gotham web application.

The Router is capable of delegating requests to secondary Router instances, which allows the support of โmodular applicationsโ. A modular application contains multiple applications within a single binary that have clear boundaries established via Rust module separation. Please see the documentation for DrawRoutes::delegate within gotham::router::builder in order to delegate to other Router instances.